IBSng accounting
IBSng II Lan Accounting
IBSng Lan Accounting is a system for policymaking, management, monitoring, and control of network resource consumption (Internet and local and enterprise servers) in organizations, offices, and companies. This system helps the managers of the communication and information technology department of the organization to optimize the use of network resources in addition to monitoring the network users. The use of Lan Accounting system helps to increase the efficiency of the people in the organization, and by controlling the accesses, it also increases the security of the system, which will lead to a reduction in costs by optimizing the network.

IBSNG is a powerful system that, relying on 20 years of presence in more than 3000 networks across the country in different dimensions and different needs, can provide all your needs to control and maintain the maximum quality of the network.